[TASK] Improve removing or records during incremental indexing
During incremental tt_content record which
should not be indexed (e.g. parent page has
"no_search" flag is set) will
now be removed from the index.

Additionally instantiating indexerStatusService
and indexRepository has now been moved to the
IndexerBase class, along with the functions
removeRecordFromIndex, removeFileFromIndex and
detectFileLanguage.
